requirements.txt
torch==1.13.1
transformers==4.24.0
ray[default]==2.0.0
mlflow==1.29.0
pyyaml==6.0
tqdm==4.64.1
fastapi==0.85.1
uvicorn==0.18.3
wandb==0.13.4
pytest==7.1.3
jsonschema==4.16.0